+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### New features
 
 - [#3262](https://github.com/clojure-emacs/cider/issues/3262) Add navigation functionality to `npfb` keys inside inspect buffer.
+### Changes
+
+- Allow using `npx nbb` as `cider-nbb-command`.
+- [#3281](https://github.com/clojure-emacs/cider/pull/3281): Replace newline chars with actual newlines in `*cider-test-report*` buffer, for prettier error messages.
 
 ## 1.6.0 (2022-12-21)